+++ cmd/ftep/main.go
@@ -47,7 +47,7 @@ func newRootCommand(out io.Writer, errOut io.Writer) *cobra.Command {
 	cmd.SetOut(out)
 	cmd.SetErr(errOut)
 	cmd.SetVersionTemplate("{{.Version}}\n")
-	cmd.AddCommand(newSearchCommand(), newReadsCommand(), newGetFieldsCommand())
+	cmd.AddCommand(newSearchCommand(), newReadsCommand(), newOpenCommand(), newGetFieldsCommand())
 	return cmd
 }